+++ src/Monitor/Monitor/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Added parameter `Dimension` for `Get-AzMetric` to easily filter metrics by dimensions [#22320]
 
 ## Version 4.5.0
 * Added cmdlets for monitor workspace: